requirements.txt
appdirs==1.4.4
APScheduler==3.10.1
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
Brotli==1.0.9
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
cssselect2==0.6.0
Django==4.1
django-extensions==3.2.0
django-psdb-engine==1.0.6
fasteners==0.18
fonttools==4.37.4
future==0.18.2
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.57.0
graphviz==0.20.1
html5lib==1.1
httplib2==0.21.0
idna==3.3
importlib-metadata==5.0.0
lxml==4.9.1
mysql-connector-python==8.0.30
mysqlclient
oauth2client==4.1.3
oauthlib==3.2.2
oscrypto==1.3.0
Pillow==9.2.0
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
pydyf==0.5.0
pyee==8.2.2
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
pyOpenSSL==22.1.0
pyparsing==3.0.9
PyPDF2==1.28.6
PyPDF3==1.0.6
pyphen==0.13.0
pyppeteer==0.2.6
python-bidi==0.4.2
python-docx==0.8.11
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.11
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
svglib==1.4.1
syncasync==20180812
tinycss2==1.1.1
tqdm==4.64.1
tzdata==2022.2
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.12
weasyprint==57.0
webencodings==0.5.1
websockets==10.3
zipp==3.9.0
zopfli==0.2.1
